SVN r3835 and r3836
Rework ListMidi so it can be more easily added to more backends.
