+++ include/Makefile.am
@@ -21,6 +21,7 @@ keyboard.h \
 logging.h \
 mapper.h \
 mem.h \
+midi.h \
 mixer.h \
 modules.h \
 mouse.h \
