+++ src/hardware/mixer.cpp
@@ -57,6 +57,7 @@
 #include "mapper.h"
 #include "hardware.h"
 #include "programs.h"
+#include "midi.h"
 
 #define MIXER_SSIZE 4
 #define MIXER_VOLSHIFT 13
@@ -871,17 +872,8 @@ private:
     }
 
     void ListMidi(){
-#if defined (WIN32)
-        unsigned int total = midiOutGetNumDevs();   
-        for(unsigned int i=0;i<total;i++) {
-            MIDIOUTCAPS mididev;
-            midiOutGetDevCaps(i, &mididev, sizeof(MIDIOUTCAPS));
-            WriteOut("%2d\t \"%s\"\n",i,mididev.szPname);
-        }
-#endif
-    return;
+        if(midi.handler) midi.handler->ListAll(this);
     };
-
 };
 
 static void MIXER_ProgramStart(Program * * make) {
